removed wagmi config and exports and moved code to wagmi provider component
